requirements.txt
streamlit==1.39.0
databricks-sql-connector==3.4.0
pandas>=2.0,<3.0
requests>=2.31.0
pyarrow>=12
